Move the custom inflector rules into a config folder
It's not a model class so it's confusing that it was stored in the models folder. This also allows us to remove the export default hack.
