+++ app/app.js
@@ -2,7 +2,7 @@ import Application from '@ember/application';
 import Resolver from 'ember-resolver';
 import loadInitializers from 'ember-load-initializers';
 import config from 'frontend-dashboard/config/environment';
-import './models/custom-inflector-rules';
+import './config/custom-inflector-rules';
 
 export default class App extends Application {
   modulePrefix = config.modulePrefix;
